pom verkleinert, baseEntity in der Api-Lib zurückgesetzt
